+++ src/renderer/src/app/core/routes/app-routes.const.ts
@@ -1,12 +1,14 @@
 const PATHS = {
-  home:    '',
-  clients: 'clients',
+  home:            '',
+  clients:         'clients',
+  settingsCompany: 'settings/company',
 } as const
 
 export const AppRoutes = {
   paths: PATHS,
   nav: {
-    home:    `/${PATHS.home}`,
-    clients: `/${PATHS.clients}`,
+    home:            `/${PATHS.home}`,
+    clients:         `/${PATHS.clients}`,
+    settingsCompany: `/${PATHS.settingsCompany}`,
   },
 } as const