+++ src/renderer/src/app/stores/company/company-store.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { CompanyStore } from './company-store';
+
+describe('CompanyStore', () => {
+  let service: CompanyStore;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(CompanyStore);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});